bookbook_check/BOOK_PJT/BACKEND/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.1
certifi==2025.11.12
colorama==0.4.6
distro==1.9.0
Django==5.2.4
django-cors-headers==4.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.3.5
openai==2.11.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
